refactor(steps-6-7): extract SynthesisPipeline and MusicScheduler
Step 6: SynthesisPipeline class in voxd/synthesis.py (464 lines) with
synthesize_to_file, try_direct_play, vibe/model helpers. Cache and
mutex injected via constructor. 8 test classes moved.

Step 7: MusicScheduler class in voxd/music_scheduler.py (533 lines)
with loop, kill_proc, 10 music fields moved off DaemonContext with
delegation properties. 6 test classes moved.

_monolith.py: 1645→1121 lines (-524). 8 metrics improved:
method_ratio +0.318, encapsulation_ratio +0.166, avg_params -0.720,
max_complexity 20→13, avg_complexity -1.690, module_size -524,
classes_per_module 2→1, public_attr_violations 17→7.
Test count: 1466→1467.
