+++ .oo-audit.jsonl
@@ -20,3 +20,7 @@
 {"ts":"2026-05-14T01:31:26Z","commit":"0b49063","files_scored":41,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089],"encapsulation_ratio":[0.0,0.056],"avg_params":[2.23,2.43],"avg_complexity":[3.88,3.71],"module_size":[1922.0,1645.0],"classes_per_module":[3.0,2.0],"public_attr_violations":[19.0,17.0]}}}
 {"ts":"2026-05-14T01:34:33Z","commit":"0b49063","files_scored":41,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089]}}}
 {"ts":"2026-05-14T01:38:05Z","commit":"0b49063","files_scored":41,"files_improved":4,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[110.0,114.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089],"encapsulation_ratio":[0.0,0.056],"avg_params":[2.23,2.43],"avg_complexity":[3.88,3.71],"module_size":[1922.0,1645.0],"classes_per_module":[3.0,2.0],"public_attr_violations":[19.0,17.0]},"src/punt_vox/voxd/playback.py":{"method_ratio":[0.0,0.5],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.67],"max_complexity":[0.0,13.0],"avg_complexity":[0.0,2.22],"module_size":[0.0,346.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,0.182],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/track_generator.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.33],"max_complexity":[0.0,5.0],"avg_complexity":[0.0,2.33],"module_size":[0.0,73.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T02:01:26Z","commit":"4a552e5","files_scored":43,"files_improved":2,"files_regressed":2,"verdict":"fail","deltas":{"src/punt_vox/voxd/music_scheduler.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.61],"max_complexity":[0.0,20.0],"avg_complexity":[0.0,2.36],"module_size":[0.0,464.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/synthesis.py":{"method_ratio":[0.0,0.714],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,5.75],"max_complexity":[0.0,15.0],"avg_complexity":[0.0,5.12],"module_size":[0.0,407.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,0.333],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T02:03:36Z","commit":"4a552e5","files_scored":43,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.089,0.407],"encapsulation_ratio":[0.056,0.222],"avg_params":[2.43,1.71],"max_complexity":[20.0,13.0],"avg_complexity":[3.71,2.02],"module_size":[1645.0,1121.0],"classes_per_module":[2.0,1.0],"public_attr_violations":[17.0,7.0]}}}
+{"ts":"2026-05-14T02:04:47Z","commit":"4a552e5","files_scored":43,"files_improved":2,"files_regressed":2,"verdict":"fail","deltas":{"src/punt_vox/voxd/music_scheduler.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.61],"max_complexity":[0.0,20.0],"avg_complexity":[0.0,2.36],"module_size":[0.0,464.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/synthesis.py":{"method_ratio":[0.0,0.714],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,5.75],"max_complexity":[0.0,15.0],"avg_complexity":[0.0,5.12],"module_size":[0.0,407.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,0.333],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T02:06:41Z","commit":"4a552e5","files_scored":43,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.089,0.407],"encapsulation_ratio":[0.056,0.222],"avg_params":[2.43,1.71],"max_complexity":[20.0,13.0],"avg_complexity":[3.71,2.02],"module_size":[1645.0,1121.0],"classes_per_module":[2.0,1.0],"public_attr_violations":[17.0,7.0]}}}
