+++ .oo-baseline.json
@@ -434,7 +434,7 @@
     "avg_params": 0.0,
     "max_complexity": 0.0,
     "avg_complexity": 0.0,
-    "module_size": 114.0,
+    "module_size": 134.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -442,16 +442,16 @@
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/_monolith.py": {
-    "method_ratio": 0.089,
-    "encapsulation_ratio": 0.056,
-    "avg_params": 2.43,
-    "max_complexity": 20.0,
-    "avg_complexity": 3.71,
-    "module_size": 1645.0,
-    "classes_per_module": 2.0,
-    "class_to_func_ratio": 0.047,
+    "method_ratio": 0.407,
+    "encapsulation_ratio": 0.222,
+    "avg_params": 1.71,
+    "max_complexity": 13.0,
+    "avg_complexity": 2.02,
+    "module_size": 1121.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 0.028,
     "init_violations": 1.0,
-    "public_attr_violations": 17.0,
+    "public_attr_violations": 7.0,
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/chimes.py": {
@@ -493,6 +493,19 @@
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
   },
+  "src/punt_vox/voxd/music_scheduler.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.61,
+    "max_complexity": 20.0,
+    "avg_complexity": 2.36,
+    "module_size": 464.0,
+    "classes_per_module": 2.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
   "src/punt_vox/voxd/playback.py": {
     "method_ratio": 0.5,
     "encapsulation_ratio": 1.0,
@@ -502,6 +515,19 @@
     "module_size": 346.0,
     "classes_per_module": 2.0,
     "class_to_func_ratio": 0.182,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/synthesis.py": {
+    "method_ratio": 0.714,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 5.75,
+    "max_complexity": 15.0,
+    "avg_complexity": 5.12,
+    "module_size": 407.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 0.333,
     "init_violations": 0.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
