+++ tests/test_voxd.py
@@ -4,20 +4,16 @@
 from __future__ import annotations
 
 import asyncio
-import contextlib
 import logging
-import os
 from pathlib import Path
 from typing import cast
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from conftest import _get_valid_mp3_bytes  # pyright: ignore[reportPrivateUsage]
 
 from punt_vox.voxd import (
     DaemonContext,
     PlaybackItem,
-    _apply_vibe_for_synthesis,
     _auto_track_name,
     _handle_music_list,
     _handle_music_next,
@@ -29,10 +25,6 @@ from punt_vox.voxd import (
     _health_payload_full,
     _health_payload_minimal,
     _health_route,
-    _kill_music_proc,
-    _model_supports_expressive_tags,
-    _music_loop,
-    _try_direct_play,
 )
 
 
@@ -197,171 +189,6 @@ class TestHealthPayloadMinimal:
         assert "secret stderr" not in raw.decode()
 
 
-class TestTryDirectPlay:
-    """Voxd dispatches to provider.play_directly for local providers."""
-
-    def _run(self, provider: MagicMock, ctx: DaemonContext) -> int | None | Exception:
-        with patch("punt_vox.voxd._monolith.get_provider", return_value=provider):
-            return asyncio.run(
-                _try_direct_play(
-                    text="hello",
-                    voice=None,
-                    provider_name="espeak",
-                    model=None,
-                    language=None,
-                    rate=None,
-                    vibe_tags=None,
-                    stability=None,
-                    similarity=None,
-                    style=None,
-                    speaker_boost=None,
-                    api_key=None,
-                    ctx=ctx,
-                )
-            )
-
-    def test_returns_provider_rc_on_success(self) -> None:
-        ctx = _make_ctx()
-        provider = MagicMock()
-        provider.play_directly = MagicMock(return_value=0)
-
-        rc = self._run(provider, ctx)
-
-        assert rc == 0
-        assert ctx.last_playback is not None
-        assert ctx.last_playback["rc"] == 0
-        provider.play_directly.assert_called_once()
-
-    def test_returns_none_for_cloud_provider(self) -> None:
-        """A provider lacking play_directly opts out of the direct-play path."""
-        ctx = _make_ctx()
-        # spec= without play_directly means hasattr/isinstance both fail.
-        provider = MagicMock(spec=["name", "synthesize"])
-
-        rc = self._run(provider, ctx)
-
-        assert rc is None
-        # No playback attempted -- last_playback stays None.
-        assert ctx.last_playback is None
-
-    def test_nonzero_rc_logs_error(self, caplog: pytest.LogCaptureFixture) -> None:
-        ctx = _make_ctx()
-        provider = MagicMock()
-        provider.play_directly = MagicMock(return_value=2)
-
-        with caplog.at_level(logging.ERROR, logger="punt_vox.voxd"):
-            rc = self._run(provider, ctx)
-
-        assert rc == 2
-        assert "Direct-play FAILED" in caplog.text
-        assert ctx.last_playback is not None
-        assert ctx.last_playback["rc"] == 2
-
-    def test_get_provider_exception_returned(
-        self, caplog: pytest.LogCaptureFixture
-    ) -> None:
-        """Provider construction failure surfaces as an Exception, not a crash."""
-        ctx = _make_ctx()
-
-        with (
-            caplog.at_level(logging.ERROR, logger="punt_vox.voxd"),
-            patch(
-                "punt_vox.voxd._monolith.get_provider",
-                side_effect=ValueError("unknown provider"),
-            ),
-        ):
-            result = asyncio.run(
-                _try_direct_play(
-                    text="hello",
-                    voice=None,
-                    provider_name="espeak",
-                    model=None,
-                    language=None,
-                    rate=None,
-                    vibe_tags=None,
-                    stability=None,
-                    similarity=None,
-                    style=None,
-                    speaker_boost=None,
-                    api_key=None,
-                    ctx=ctx,
-                )
-            )
-
-        assert isinstance(result, ValueError)
-        assert "unknown provider" in str(result)
-        assert "Direct-play raised" in caplog.text
-
-    def test_no_api_key_skips_env_lock(self) -> None:
-        """Local providers without an API key must not block on _env_lock."""
-        ctx = _make_ctx()
-        provider = MagicMock()
-        provider.play_directly = MagicMock(return_value=0)
-
-        # Replace _env_lock with one that records every acquire attempt.
-        sentinel_lock = MagicMock(wraps=asyncio.Lock())
-        sentinel_lock.__aenter__ = AsyncMock()
-        sentinel_lock.__aexit__ = AsyncMock(return_value=None)
-
-        with (
-            patch("punt_vox.voxd._monolith.get_provider", return_value=provider),
-            patch("punt_vox.voxd._monolith._env_lock", sentinel_lock),
-        ):
-            asyncio.run(
-                _try_direct_play(
-                    text="hello",
-                    voice=None,
-                    provider_name="espeak",
-                    model=None,
-                    language=None,
-                    rate=None,
-                    vibe_tags=None,
-                    stability=None,
-                    similarity=None,
-                    style=None,
-                    speaker_boost=None,
-                    api_key=None,
-                    ctx=ctx,
-                )
-            )
-
-        sentinel_lock.__aenter__.assert_not_called()
-
-    def test_api_key_acquires_env_lock_for_cloud_provider(self) -> None:
-        """API-key path must serialize via _env_lock to protect os.environ."""
-        ctx = _make_ctx()
-        provider = MagicMock()
-        provider.play_directly = MagicMock(return_value=0)
-
-        sentinel_lock = MagicMock()
-        sentinel_lock.__aenter__ = AsyncMock()
-        sentinel_lock.__aexit__ = AsyncMock(return_value=None)
-
-        with (
-            patch("punt_vox.voxd._monolith.get_provider", return_value=provider),
-            patch("punt_vox.voxd._monolith._env_lock", sentinel_lock),
-        ):
-            asyncio.run(
-                _try_direct_play(
-                    text="hello",
-                    voice=None,
-                    provider_name="elevenlabs",
-                    model=None,
-                    language=None,
-                    rate=None,
-                    vibe_tags=None,
-                    stability=None,
-                    similarity=None,
-                    style=None,
-                    speaker_boost=None,
-                    api_key="secret",
-                    ctx=ctx,
-                )
-            )
-
-        sentinel_lock.__aenter__.assert_called_once()
-
-
 class TestHandleSynthesizeShortCircuit:
     """``_handle_synthesize`` skips _try_direct_play for cloud providers."""
 
@@ -406,1013 +233,6 @@ class TestHandleSynthesizeShortCircuit:
         mock_direct.assert_called_once()
         call_kwargs = mock_direct.call_args.kwargs
         assert call_kwargs["provider_name"] == "espeak"
-
-
-class TestDirectPlayProtocol:
-    """Direct-play capability is detected via the DirectPlayProvider protocol."""
-
-    def test_polly_is_not_direct_play(self) -> None:
-        from punt_vox.providers.polly import PollyProvider
-        from punt_vox.types import DirectPlayProvider
-
-        with patch("punt_vox.providers.polly.boto3.client"):
-            provider = PollyProvider()
-        assert not isinstance(provider, DirectPlayProvider)
-
-    def test_openai_is_not_direct_play(self) -> None:
-        from punt_vox.providers.openai import OpenAIProvider
-        from punt_vox.types import DirectPlayProvider
-
-        with patch.dict("os.environ", {"OPENAI_API_KEY": "sk-test"}):
-            provider = OpenAIProvider()
-        assert not isinstance(provider, DirectPlayProvider)
-
-    def test_elevenlabs_is_not_direct_play(self) -> None:
-        from punt_vox.providers.elevenlabs import ElevenLabsProvider
-        from punt_vox.types import DirectPlayProvider
-
-        with patch.dict("os.environ", {"ELEVENLABS_API_KEY": "test"}):
-            provider = ElevenLabsProvider()
-        assert not isinstance(provider, DirectPlayProvider)
-
-    def test_espeak_is_direct_play(self) -> None:
-        from punt_vox.providers.espeak import EspeakProvider
-        from punt_vox.types import DirectPlayProvider
-
-        with patch(
-            "punt_vox.providers.espeak._find_espeak_binary",
-            return_value="/usr/bin/espeak-ng",
-        ):
-            provider = EspeakProvider()
-        assert isinstance(provider, DirectPlayProvider)
-
-
-class TestApiKeyPassthroughIntegration:
-    """End-to-end api_key flow: WebSocket -> _handle_record -> provider.
-
-    Verification of vox-a3e: a single user maintaining multiple provider
-    keys (key-A for billing project alpha, key-B for billing project
-    bravo) can scope a single synthesize/record call to a specific key
-    without leaking that key into the process environment beyond the
-    single request, and without cross-call contamination.
-
-    The test drives the real Starlette app via ``starlette.testclient``,
-    sends record messages over a real WebSocket, and registers a stub
-    provider via ``punt_vox.voxd.get_provider``. The stub inspects
-    ``os.environ[ELEVENLABS_API_KEY]`` at the moment the factory is
-    called — that is the exact point where the per-call key is visible
-    to the provider code. Recording the observed value proves the
-    passthrough is working and isolated per call.
-    """
-
-    def _build_stub_provider(self, observed: list[str | None]) -> type:
-        """Return a stub TTSProvider class that records the observed api key."""
-        from punt_vox.types import (
-            AudioProviderId,
-            AudioRequest,
-            AudioResult,
-            HealthCheck,
-        )
-
-        valid_mp3 = _get_valid_mp3_bytes()
-
-        class _StubProvider:
-            name = "elevenlabs"
-            default_voice = "matilda"
-            supports_expressive_tags = False
-
-            def __init__(self) -> None:
-                # Read env at factory time — this is the point at which
-                # voxd's ``_synthesize_to_file`` has mutated os.environ
-                # for a per-call key. Record whatever is there (including
-                # None when the caller did not pass ``api_key``).
-                observed.append(os.environ.get("ELEVENLABS_API_KEY"))
-
-            def synthesize(
-                self, request: AudioRequest, output_path: Path
-            ) -> AudioResult:
-                output_path.write_bytes(valid_mp3)
-                return AudioResult(
-                    path=output_path,
-                    text=request.text,
-                    provider=AudioProviderId.elevenlabs,
-                    voice="matilda",
-                )
-
-            def generate_audio(self, request: AudioRequest) -> AudioResult:
-                raise NotImplementedError
-
-            def generate_audios(self, requests: object) -> list[AudioResult]:
-                raise NotImplementedError
-
-            def resolve_voice(self, name: str, language: str | None = None) -> str:
-                return name or "matilda"
-
-            def get_default_voice(self, language: str) -> str:
-                return "matilda"
-
-            def list_voices(self, language: str | None = None) -> list[str]:
-                return ["matilda"]
-
-            def infer_language_from_voice(self, voice: str) -> str | None:
-                return None
-
-            def check_health(self) -> list[HealthCheck]:
-                return []
-
-        return _StubProvider
-
-    def _run_record(
-        self,
-        api_key: str | None,
-        observed: list[str | None],
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> dict[str, object]:
-        """Send a record message over a real WebSocket and return the response."""
-        from starlette.testclient import TestClient
-
-        from punt_vox.voxd import build_app
-
-        stub_provider_cls = self._build_stub_provider(observed)
-
-        def fake_get_provider(
-            name: str,
-            *,
-            config_dir: object = None,
-            model: str | None = None,
-        ) -> object:
-            assert name == "elevenlabs"
-            return stub_provider_cls()
-
-        monkeypatch.setattr("punt_vox.voxd._monolith.get_provider", fake_get_provider)
-
-        # Disable the cache so every call reaches the stub factory.
-        def _cache_miss(
-            _text: str,
-            _voice: str,
-            _provider: str,
-            _api_key: str | None = None,
-        ) -> Path | None:
-            return None
-
-        def _cache_noop(
-            _text: str,
-            _voice: str,
-            _provider: str,
-            _path: Path,
-            _api_key: str | None = None,
-        ) -> None:
-            return None
-
-        monkeypatch.setattr("punt_vox.voxd._monolith.cache_get", _cache_miss)
-        monkeypatch.setattr("punt_vox.voxd._monolith.cache_put", _cache_noop)
-
-        ctx = DaemonContext(auth_token=None, port=0)
-        app = build_app(ctx)
-
-        msg: dict[str, object] = {
-            "type": "record",
-            "id": "test-rec",
-            "text": "billable synthesis",
-            "provider": "elevenlabs",
-            "voice": "matilda",
-        }
-        if api_key is not None:
-            msg["api_key"] = api_key
-
-        with (
-            TestClient(app) as client,
-            client.websocket_connect("/ws") as ws,
-        ):
-            ws.send_json(msg)
-            response: dict[str, object] = ws.receive_json()
-
-        return response
-
-    def test_first_call_key_reaches_provider(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Sending ``api_key=key-alpha`` exposes exactly that key to the provider."""
-        # Clear any ambient ELEVENLABS_API_KEY so the baseline is well-defined.
-        monkeypatch.delenv("ELEVENLABS_API_KEY", raising=False)
-        observed: list[str | None] = []
-
-        response = self._run_record("test-key-alpha-billing", observed, monkeypatch)
-
-        assert response.get("type") == "audio"
-        assert len(observed) == 1
-        assert observed[0] == "test-key-alpha-billing"
-
-    def test_second_call_key_does_not_leak_from_first(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Two sequential calls with different keys: each provider sees its own.
-
-        This is the load-bearing invariant for single-user multi-key
-        billing isolation. If voxd leaked key-alpha into the second
-        call, the user would be billed under the wrong project.
-        """
-        monkeypatch.delenv("ELEVENLABS_API_KEY", raising=False)
-        observed: list[str | None] = []
-
-        # Call 1 — key alpha
-        self._run_record("test-key-alpha-billing", observed, monkeypatch)
-        # Call 2 — key bravo (fresh app instance, but shares os.environ)
-        self._run_record("test-key-bravo-billing", observed, monkeypatch)
-
-        assert observed == [
-            "test-key-alpha-billing",
-            "test-key-bravo-billing",
-        ]
-        # Neither key leaked back into the process environment after
-        # the calls returned.
-        assert os.environ.get("ELEVENLABS_API_KEY") is None
-
-    def test_no_api_key_falls_back_to_ambient_env(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """``api_key=None`` leaves whatever is already in os.environ alone.
-
-        This is the keys.env-fallback path: when the caller does not
-        pass a per-call key, voxd does not touch ``os.environ``, so
-        the provider factory sees whatever the keys.env loader put
-        there at daemon startup (or whatever the current test ambient
-        env has). The key invariant is that no cross-call mutation
-        happens.
-        """
-        # Set a sentinel ambient key — the provider should see it.
-        monkeypatch.setenv("ELEVENLABS_API_KEY", "ambient-fallback-key")
-        observed: list[str | None] = []
-
-        self._run_record(None, observed, monkeypatch)
-
-        assert len(observed) == 1
-        assert observed[0] == "ambient-fallback-key"
-        # Ambient key still present — no cross-call mutation.
-        assert os.environ.get("ELEVENLABS_API_KEY") == "ambient-fallback-key"
-
-    def test_previously_set_ambient_key_restored_after_per_call_key(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Per-call key is restored to the ambient value after the call.
-
-        voxd's ``_synthesize_to_file`` saves the old env value, sets
-        the per-call key, runs the provider, then restores the old
-        value in a ``finally`` block. This test verifies the restore.
-        """
-        monkeypatch.setenv("ELEVENLABS_API_KEY", "ambient-persistent-key")
-        observed: list[str | None] = []
-
-        self._run_record("per-call-override", observed, monkeypatch)
-
-        assert observed == ["per-call-override"]
-        # After the call, the original ambient value is back.
-        assert os.environ.get("ELEVENLABS_API_KEY") == "ambient-persistent-key"
-
-
-class TestCacheApiKeyBypass:
-    """Cache bypass when ``api_key`` is set — vox-a3e billing isolation.
-
-    The round-3 design tried to *partition* the cache by ``api_key`` by
-    folding the key into the digest. CodeQL's
-    ``py/weak-sensitive-data-hashing`` rule correctly flagged that: a
-    regular cryptographic hash is inappropriate for hashing password-
-    class material even for a filename-only use case, and the only
-    lint-clean alternatives are password KDFs (Argon2, scrypt, bcrypt,
-    PBKDF2) whose per-call cost is unacceptable for cache filenames.
-
-    Round 4 switches to **cache bypass**: per-call api_key calls skip
-    ``cache_get`` and ``cache_put`` entirely at the voxd call site,
-    so no sensitive data ever reaches ``cache.py``. The correctness
-    story also improves — a billing-isolated call can never read bytes
-    synthesized under a different key, and can never leave bytes
-    behind that a different key could later reuse.
-
-    These tests drive the real voxd WebSocket code path. The bypass
-    tests use spies on ``cache_get``/``cache_put`` to prove the
-    bypass-not-call behavior directly. The no-poison test uses the
-    real on-disk cache (``cache.py``'s ``cache_get``/``cache_put``
-    untouched, ``CACHE_DIR`` pointed at a tmp dir) to prove that an
-    api_key synthesis does not disturb the anonymous cache entry for
-    the same text. The provider-distinct test is inherited from the
-    round-3 suite and locks in that per-call api_key values still
-    reach the provider correctly.
-    """
-
-    def _build_counting_provider(self, calls: list[str | None]) -> type:
-        """Stub provider that appends to ``calls`` on every construction."""
-        from punt_vox.types import (
-            AudioProviderId,
-            AudioRequest,
-            AudioResult,
-            HealthCheck,
-        )
-
-        valid_mp3 = _get_valid_mp3_bytes()
-
-        class _CountingProvider:
-            name = "elevenlabs"
-            default_voice = "matilda"
-            supports_expressive_tags = False
-
-            def __init__(self) -> None:
-                # Record the api_key visible at construction time. The
-                # key invariant is that the provider factory runs once
-                # per distinct api_key even when text/voice/provider
-                # are identical across calls.
-                calls.append(os.environ.get("ELEVENLABS_API_KEY"))
-
-            def synthesize(
-                self, request: AudioRequest, output_path: Path
-            ) -> AudioResult:
-                output_path.write_bytes(valid_mp3)
-                return AudioResult(
-                    path=output_path,
-                    text=request.text,
-                    provider=AudioProviderId.elevenlabs,
-                    voice="matilda",
-                )
-
-            def generate_audio(self, request: AudioRequest) -> AudioResult:
-                raise NotImplementedError
-
-            def generate_audios(self, requests: object) -> list[AudioResult]:
-                raise NotImplementedError
-
-            def resolve_voice(self, name: str, language: str | None = None) -> str:
-                return name or "matilda"
-
-            def get_default_voice(self, language: str) -> str:
-                return "matilda"
-
-            def list_voices(self, language: str | None = None) -> list[str]:
-                return ["matilda"]
-
-            def infer_language_from_voice(self, voice: str) -> str | None:
-                return None
-
-            def check_health(self) -> list[HealthCheck]:
-                return []
-
-        return _CountingProvider
-
-    def _run_record(
-        self,
-        *,
-        api_key: str | None,
-        text: str,
-        calls: list[str | None],
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> dict[str, object]:
-        """Send a ``record`` message over a real WebSocket and return the response.
-
-        Does NOT patch ``cache_get``/``cache_put`` — the caller is
-        responsible for patching them (with spies or with a tmp
-        ``CACHE_DIR``) before invoking this helper.
-        """
-        from starlette.testclient import TestClient
-
-        from punt_vox.voxd import build_app
-
-        provider_cls = self._build_counting_provider(calls)
-
-        def fake_get_provider(
-            name: str,
-            *,
-            config_dir: object = None,
-            model: str | None = None,
-        ) -> object:
-            assert name == "elevenlabs"
-            return provider_cls()
-
-        monkeypatch.setattr("punt_vox.voxd._monolith.get_provider", fake_get_provider)
-
-        ctx = DaemonContext(auth_token=None, port=0)
-        app = build_app(ctx)
-
-        msg: dict[str, object] = {
-            "type": "record",
-            "id": "test-rec",
-            "text": text,
-            "provider": "elevenlabs",
-            "voice": "matilda",
-        }
-        if api_key is not None:
-            msg["api_key"] = api_key
-
-        with (
-            TestClient(app) as client,
-            client.websocket_connect("/ws") as ws,
-        ):
-            ws.send_json(msg)
-            response: dict[str, object] = ws.receive_json()
-
-        return response
-
-    def test_api_key_call_bypasses_cache_get_and_put(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        tmp_path: Path,
-    ) -> None:
-        """Per-call api_key synthesis must not call cache_get or cache_put.
-
-        Spy-based assertion (not mock-return-value): we replace the two
-        cache entry points with functions that record their call args
-        in a list, then drive the real voxd WebSocket handler twice —
-        once anonymously, once with an api_key. The anonymous call
-        MUST hit both spies; the api_key call MUST hit neither.
-
-        Spies record actual behavior directly. A mock-return-value
-        approach would only prove the code handled whatever fake
-        return value the test injected, which does not distinguish
-        "the bypass works" from "cache_get was called and returned
-        None".
-        """
-        # Real CACHE_DIR pointed at tmp so the backing cache.py code
-        # never touches user state if a spy somehow lets the real
-        # functions through.
-        monkeypatch.setattr("punt_vox.cache.CACHE_DIR", tmp_path / "cache")
-        monkeypatch.setenv("ELEVENLABS_API_KEY", "")
-
-        get_calls: list[tuple[object, ...]] = []
-        put_calls: list[tuple[object, ...]] = []
-
-        def spy_cache_get(*args: object) -> Path | None:
-            get_calls.append(args)
-            return None  # Never hit — force synthesis on the anon path.
-
-        def spy_cache_put(*args: object) -> Path | None:
-            put_calls.append(args)
-            # Return the source path unchanged so the handler keeps
-            # using it for the record response.
-            source = args[3]
-            assert isinstance(source, Path)
-            return source
-
-        monkeypatch.setattr("punt_vox.voxd._monolith.cache_get", spy_cache_get)
-        monkeypatch.setattr("punt_vox.voxd._monolith.cache_put", spy_cache_put)
-
-        calls: list[str | None] = []
-
-        # Call 1: anonymous — both spies MUST fire.
-        resp1 = self._run_record(
-            api_key=None,
-            text="bypass test phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp1.get("type") == "audio"
-        assert len(get_calls) == 1, (
-            f"anonymous call: expected cache_get spy to fire once, "
-            f"got {len(get_calls)}: {get_calls}"
-        )
-        assert len(put_calls) == 1, (
-            f"anonymous call: expected cache_put spy to fire once, "
-            f"got {len(put_calls)}: {put_calls}"
-        )
-
-        # Call 2: api_key set — NEITHER spy may fire. The cache_get /
-        # cache_put call counts stay locked at 1 (from call 1).
-        resp2 = self._run_record(
-            api_key="sk_bypass_test",
-            text="bypass test phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp2.get("type") == "audio"
-        assert len(get_calls) == 1, (
-            f"api_key call must bypass cache_get, got {len(get_calls)} "
-            f"total calls: {get_calls}"
-        )
-        assert len(put_calls) == 1, (
-            f"api_key call must bypass cache_put, got {len(put_calls)} "
-            f"total calls: {put_calls}"
-        )
-        # And the provider was invoked twice total (once per call),
-        # proving the api_key call actually synthesized rather than
-        # returning an early short-circuit.
-        assert len(calls) == 2, (
-            f"expected provider called twice (once anon, once api_key), "
-            f"got {len(calls)}: {calls}"
-        )
-
-    def test_api_key_call_does_not_poison_anonymous_cache(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        tmp_path: Path,
-    ) -> None:
-        """An api_key synthesis must not disturb the anonymous cache entry.
-
-        Sequence:
-          1. Anonymous call — provider runs, writes cache entry for
-             the (text, voice, provider) tuple.
-          2. Anonymous call (same text) — cache hit, provider NOT
-             called again.
-          3. api_key call (same text) — cache BYPASSED, provider IS
-             called (proving bypass-not-read).
-          4. Anonymous call (same text) — cache hit from step 1 is
-             STILL REACHABLE (proving bypass-not-write: the api_key
-             call did not overwrite or evict the anon entry).
-
-        Uses the real on-disk cache module (not mocked) with
-        ``CACHE_DIR`` redirected to a tmp dir.
-        """
-        monkeypatch.delenv("ELEVENLABS_API_KEY", raising=False)
-        monkeypatch.setattr("punt_vox.cache.CACHE_DIR", tmp_path / "cache")
-        calls: list[str | None] = []
-
-        # Step 1: anonymous, cold cache — provider runs.
-        resp1 = self._run_record(
-            api_key=None,
-            text="shared anon phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp1.get("type") == "audio"
-        assert len(calls) == 1, f"step 1 should synthesize, got calls={calls}"
-
-        # Step 2: anonymous again — cache hit, no provider call.
-        resp2 = self._run_record(
-            api_key=None,
-            text="shared anon phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp2.get("type") == "audio"
-        assert len(calls) == 1, (
-            f"step 2 should hit cache (no new provider call), got calls={calls}"
-        )
-
-        # Step 3: api_key set — cache BYPASSED, provider runs again.
-        resp3 = self._run_record(
-            api_key="sk_poison_test",
-            text="shared anon phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp3.get("type") == "audio"
-        assert len(calls) == 2, (
-            f"step 3: api_key call must bypass cache and synthesize, "
-            f"expected 2 provider calls total, got calls={calls}"
-        )
-
-        # Step 4: anonymous again — the original cache entry from step
-        # 1 must STILL be reachable. If the api_key call had poisoned
-        # the anon cache (e.g. overwritten the file or removed it),
-        # we would see a 3rd provider call here.
-        resp4 = self._run_record(
-            api_key=None,
-            text="shared anon phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp4.get("type") == "audio"
-        assert len(calls) == 2, (
-            f"step 4: anonymous cache entry was POISONED by the api_key "
-            f"call; anon call should still hit cache, got calls={calls}"
-        )
-
-    def test_per_call_keys_reach_provider_distinct(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        tmp_path: Path,
-    ) -> None:
-        """Two consecutive api_key calls reach the provider with the right key.
-
-        This locks in the core billing-isolation invariant: each
-        per-call api_key value is visible to the provider factory at
-        construction time, and the two values do not cross-contaminate.
-        Both calls use the same text so the assertion would fail if a
-        future refactor accidentally reintroduced a shared cache across
-        api_key scopes.
-        """
-        monkeypatch.setenv("ELEVENLABS_API_KEY", "")
-        monkeypatch.setattr("punt_vox.cache.CACHE_DIR", tmp_path / "cache")
-        calls: list[str | None] = []
-
-        resp1 = self._run_record(
-            api_key="sk_A",
-            text="billable phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp1.get("type") == "audio"
-
-        resp2 = self._run_record(
-            api_key="sk_B",
-            text="billable phrase",
-            calls=calls,
-            monkeypatch=monkeypatch,
-        )
-        assert resp2.get("type") == "audio"
-
-        # Two distinct provider constructions, each seeing its own key.
-        assert calls == ["sk_A", "sk_B"], (
-            f"expected provider constructions to see ['sk_A', 'sk_B'], got {calls}"
-        )
-
-
-class TestSynthesizeFailFast:
-    """0-byte synthesis must raise and skip cache_put to avoid poisoning."""
-
-    def test_zero_byte_output_raises_and_skips_cache(self, tmp_path: Path) -> None:
-        from punt_vox.voxd import _synthesize_to_file
-
-        # Mock client.synthesize to leave the file at 0 bytes.
-        captured_temp: dict[str, Path] = {}
-
-        def fake_synth(request: object, output_path: Path) -> None:
-            output_path.write_bytes(b"")
-            captured_temp["path"] = output_path
-
-        provider = MagicMock()
-        provider.name = "espeak"
-
-        with (
-            patch("punt_vox.voxd._monolith.get_provider", return_value=provider),
-            patch("punt_vox.voxd._monolith.cache_get", return_value=None),
-            patch("punt_vox.voxd._monolith.cache_put") as mock_cache_put,
-            patch("punt_vox.voxd._monolith.TTSClient") as mock_client_cls,
-        ):
-            mock_client = MagicMock()
-            mock_client.synthesize = fake_synth
-            mock_client_cls.return_value = mock_client
-
-            with pytest.raises(RuntimeError, match="missing or empty"):
-                asyncio.run(
-                    _synthesize_to_file(
-                        text="hello",
-                        voice=None,
-                        provider_name="espeak",
-                        model=None,
-                        language=None,
-                        rate=None,
-                        vibe_tags=None,
-                        stability=None,
-                        similarity=None,
-                        style=None,
-                        speaker_boost=None,
-                        api_key=None,
-                    )
-                )
-
-        mock_cache_put.assert_not_called()
-        # The temp file the synthesizer created must be cleaned up.
-        assert "path" in captured_temp
-        assert not captured_temp["path"].exists()
-
-
-class TestDirectPlaySerialization:
-    """Direct-play and queued playback share _playback_mutex."""
-
-    def test_concurrent_direct_plays_serialize(self) -> None:
-        """Two concurrent direct-play calls must not run in parallel."""
-        import time as time_module
-
-        ctx = _make_ctx()
-
-        starts: list[float] = []
-        ends: list[float] = []
-
-        def slow_play(_request: object) -> int:
-            starts.append(time_module.monotonic())
-            time_module.sleep(0.1)
-            ends.append(time_module.monotonic())
-            return 0
-
-        provider = MagicMock()
-        provider.play_directly = slow_play
-
-        async def _drive() -> None:
-            with patch("punt_vox.voxd._monolith.get_provider", return_value=provider):
-                await asyncio.gather(
-                    _try_direct_play(
-                        text="one",
-                        voice=None,
-                        provider_name="espeak",
-                        model=None,
-                        language=None,
-                        rate=None,
-                        vibe_tags=None,
-                        stability=None,
-                        similarity=None,
-                        style=None,
-                        speaker_boost=None,
-                        api_key=None,
-                        ctx=ctx,
-                    ),
-                    _try_direct_play(
-                        text="two",
-                        voice=None,
-                        provider_name="espeak",
-                        model=None,
-                        language=None,
-                        rate=None,
-                        vibe_tags=None,
-                        stability=None,
-                        similarity=None,
-                        style=None,
-                        speaker_boost=None,
-                        api_key=None,
-                        ctx=ctx,
-                    ),
-                )
-
-        asyncio.run(_drive())
-
-        assert len(starts) == 2
-        assert len(ends) == 2
-        # The second call must start AFTER the first call ends.
-        first_end = min(ends)
-        second_start = max(starts)
-        assert second_start >= first_end, (
-            f"direct-play calls overlapped: "
-            f"second_start={second_start}, first_end={first_end}"
-        )
-
-
-# ---------------------------------------------------------------------------
-# voxd path helpers — per-user state, not /etc or /var
-# ---------------------------------------------------------------------------
-
-
-class TestModelSupportsExpressiveTags:
-    """``_model_supports_expressive_tags`` lookup for vibe-tag gating.
-
-    Closes vox-fhl. The function must be a pure capability check that
-    runs BEFORE provider construction (so it can fire outside the
-    env-mutation lock that the real synthesize path needs).
-    """
-
-    def test_elevenlabs_v3_supported(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        assert _model_supports_expressive_tags("elevenlabs", "eleven_v3") is True
-
-    def test_elevenlabs_flash_does_not_support(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        assert (
-            _model_supports_expressive_tags("elevenlabs", "eleven_flash_v2_5") is False
-        )
-
-    def test_elevenlabs_default_supported(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """None model resolves to eleven_v3 — supports expressive tags."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        assert _model_supports_expressive_tags("elevenlabs", None) is True
-
-    def test_polly_never_supports(self) -> None:
-        assert _model_supports_expressive_tags("polly", None) is False
-        assert _model_supports_expressive_tags("polly", "neural") is False
-
-    def test_openai_never_supports(self) -> None:
-        assert _model_supports_expressive_tags("openai", None) is False
-        assert _model_supports_expressive_tags("openai", "tts-1-hd") is False
-
-    def test_say_never_supports(self) -> None:
-        assert _model_supports_expressive_tags("say", None) is False
-
-    def test_espeak_never_supports(self) -> None:
-        assert _model_supports_expressive_tags("espeak", None) is False
-
-    def test_unknown_provider_never_supports(self) -> None:
-        """Defensive default for any provider name not in the lookup."""
-        assert _model_supports_expressive_tags("future_provider", "any") is False
-
-
-class TestApplyVibeForSynthesis:
-    """``_apply_vibe_for_synthesis`` gates vibe-tag handling on capability.
-
-    Closes vox-fhl. The helper takes RAW text (NOT yet normalized) and
-    runs ``normalize_for_speech`` on the body itself, after splitting
-    leading bracket tags off. The order matters because
-    ``normalize_for_speech`` discards brackets via its non-prosody
-    punctuation filter — if normalization runs first, ``[serious]``
-    becomes ``serious`` and the bare word survives into TTS input on
-    every non-expressive provider.
-
-    With an expressive model, vibe_tags + leading user tags get
-    re-attached after normalization. With a non-expressive model,
-    both are dropped entirely.
-    """
-
-    def test_v3_preserves_vibe_tags(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """eleven_v3 interprets tags natively — vibe_tags are prepended."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "Hello world", "[excited]", "elevenlabs", "eleven_v3"
-        )
-        assert result == "[excited] Hello world"
-
-    def test_expressive_model_passthrough_when_no_vibe_tags(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "Hello world", None, "elevenlabs", "eleven_v3"
-        )
-        assert result == "Hello world"
-
-    def test_v3_preserves_user_tags_in_text(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """eleven_v3 interprets tags natively — user tags are preserved."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "[whisper] Quiet message", None, "elevenlabs", "eleven_v3"
-        )
-        assert result == "[whisper] Quiet message"
-
-    def test_non_expressive_model_drops_vibe_tags(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """When the model can't interpret tags, the vibe_tags MUST not appear."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "Hello world", "[serious]", "elevenlabs", "eleven_flash_v2_5"
-        )
-        assert result == "Hello world"
-        assert "[serious]" not in result
-        assert "serious" not in result
-
-    def test_non_expressive_model_strips_user_tags_from_text(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """User-supplied bracket tags MUST be stripped to avoid literal speech."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "[serious] Hello world", None, "elevenlabs", "eleven_flash_v2_5"
-        )
-        assert result == "Hello world"
-        assert "serious" not in result
-
-    def test_non_expressive_model_strips_user_tags_even_with_vibe_set(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Combined: user text has tags AND vibe is set; both must vanish."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "[whisper] Quiet message",
-            "[excited]",
-            "elevenlabs",
-            "eleven_flash_v2_5",
-        )
-        assert result == "Quiet message"
-        assert "whisper" not in result
-        assert "excited" not in result
-
-    def test_polly_strips_user_tags(self) -> None:
-        """Polly never supports tags — user-supplied tags must be stripped."""
-        result = _apply_vibe_for_synthesis(
-            "[serious] Important", "[calm]", "polly", None
-        )
-        assert result == "Important"
-
-    def test_openai_strips_user_tags(self) -> None:
-        result = _apply_vibe_for_synthesis("[whisper] Hello", None, "openai", "tts-1")
-        assert result == "Hello"
-
-    def test_say_strips_user_tags(self) -> None:
-        result = _apply_vibe_for_synthesis(
-            "[excited] Greetings", "[happy]", "say", None
-        )
-        assert result == "Greetings"
-
-    def test_espeak_strips_user_tags(self) -> None:
-        result = _apply_vibe_for_synthesis("[serious] Notice", None, "espeak", None)
-        assert result == "Notice"
-
-    def test_degenerate_text_only_tags_non_expressive_returns_empty(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> None:
-        """Tags-only input on a non-expressive model returns empty body.
-
-        The body after splitting leading tags is empty, so there is
-        nothing speakable. Returning the literal "[serious]" would have
-        the TTS engine speak the word "serious" (the exact bug the
-        capability gate exists to prevent). Returning empty is correct
-        — the synthesize handler upstream is responsible for catching
-        empty input as a no-op.
-        """
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "[serious]", None, "elevenlabs", "eleven_flash_v2_5"
-        )
-        assert result == ""
-
-    def test_degenerate_text_only_tags_v3_preserves(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> None:
-        """Tags-only input on eleven_v3 preserves the tag as an expressive cue."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis("[sighs]", None, "elevenlabs", "eleven_v3")
-        assert result == "[sighs]"
-
-    def test_normalizes_body_through_production_call_path(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> None:
-        """Regression test for the bug Copilot found on the first attempt.
-
-        The helper must run ``normalize_for_speech`` on the body itself
-        (so snake_case/camelCase get expanded). Earlier versions of this
-        fix called ``_apply_vibe_for_synthesis(normalize_for_speech(text))``
-        from voxd, which meant the helper saw already-normalized text
-        and ``strip_expressive_tags`` could not match the brackets that
-        normalization had already eaten. The fix moves normalization
-        INSIDE the helper, after the leading-tag split.
-        """
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        # snake_case body so we can prove normalize_for_speech ran on it.
-        result = _apply_vibe_for_synthesis(
-            "[serious] my_function works", None, "polly", None
-        )
-        # Body is normalized (underscores → spaces) AND the leading bracket
-        # tag is gone — neither as brackets nor as the bare word "serious".
-        assert result == "my function works"
-        assert "serious" not in result
-        assert "[" not in result
-        assert "_" not in result
-
-    def test_v3_normalizes_body_and_preserves_tags(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> None:
-        """eleven_v3 normalizes the body and preserves leading tags."""
-        monkeypatch.delenv("TTS_MODEL", raising=False)
-        result = _apply_vibe_for_synthesis(
-            "[whisper] my_function works", None, "elevenlabs", "eleven_v3"
-        )
-        assert result == "[whisper] my function works"
-
-    # -- vox-6ls: trailing / inline vibe tags at any position ---------------
-
-    def test_trailing_tags_stripped_non_expressive(self) -> None:
-        """Trailing tags like '[alert] [serious]' must be fully removed."""
-        result = _apply_vibe_for_synthesis(
-            "hello [alert] [serious]", None, "polly", None
-        )
-        assert result == "hello"
-        assert "alert" not in result
-        assert "serious" not in result
-
-    def test_inline_tags_stripped_non_expressive(self) -> None:
-        """Inline tags must be fully removed — no bare words left."""
-        result = _apply_vibe_for_synthesis("hello [serious] world", None, "polly", None)
-        assert result == "hello world"
-        assert "serious" not in result
-
-    def test_leading_tags_still_stripped_non_expressive(self) -> None:
-        """Leading tags continue to work (regression guard)."""
-        result = _apply_vibe_for_synthesis("[serious] hello", None, "polly", None)
-        assert result == "hello"
-        assert "serious" not in result
-
-    def _patch_expressive(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Force _model_supports_expressive_tags to return True."""
-
-        def _always_expressive(_provider: str, _model: str | None) -> bool:
-            return True
-
-        monkeypatch.setattr(
-            "punt_vox.voxd._monolith._model_supports_expressive_tags",
-            _always_expressive,
-        )
-
-    def test_trailing_tags_preserved_expressive(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Expressive model keeps trailing tags intact."""
-        self._patch_expressive(monkeypatch)
-        result = _apply_vibe_for_synthesis(
-            "hello [serious]", None, "elevenlabs", "eleven_v3"
-        )
-        assert "[serious]" in result
-        assert "hello" in result
-
-    def test_inline_tags_preserved_expressive(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Expressive model keeps inline tags intact."""
-        self._patch_expressive(monkeypatch)
-        result = _apply_vibe_for_synthesis(
-            "hello [serious] world", None, "elevenlabs", "eleven_v3"
-        )
-        assert "[serious]" in result
-        assert "hello" in result
-        assert "world" in result
-
-    def test_session_vibe_tags_prepended_expressive(
-        self, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Session vibe_tags are prepended on expressive models."""
-        self._patch_expressive(monkeypatch)
-        result = _apply_vibe_for_synthesis("hello", "[warm]", "elevenlabs", "eleven_v3")
-        assert result == "[warm] hello"
 
 
 class TestHandleSynthesizeOnceFlag:
@@ -1705,43 +525,6 @@ class TestWsRoutePeerClose:
 # ---------------------------------------------------------------------------
 # Music integration tests
 # ---------------------------------------------------------------------------
-
-
-class TestDaemonContextMusicFields:
-    """DaemonContext must have all 8 music fields with correct defaults."""
-
-    def test_music_mode_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_mode == "off"
-
-    def test_music_style_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_style == ""
-
-    def test_music_owner_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_owner == ""
-
-    def test_music_vibe_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_vibe == ("", "")
-
-    def test_music_track_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_track is None
-
-    def test_music_proc_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_proc is None
-
-    def test_music_state_default(self) -> None:
-        ctx = _make_ctx()
-        assert ctx.music_state == "idle"
-
-    def test_music_changed_default(self) -> None:
-        ctx = _make_ctx()
-        assert isinstance(ctx.music_changed, asyncio.Event)
-        assert not ctx.music_changed.is_set()
 
 
 class TestMusicHandlerRegistration:
@@ -2102,428 +885,6 @@ class TestHandleMusicNext:
         assert _HANDLERS["music_next"] is _handle_music_next
 
 
-class TestMusicLoopStateTransitions:
-    """_music_loop: generation, playback, vibe changes, crash recovery."""
-
-    def test_generates_and_plays_then_stops_on_off(self) -> None:
-        """Full cycle: mode on -> generate -> play -> mode off."""
-        ctx = _make_ctx()
-
-        call_log: list[str] = []
-
-        async def fake_generate_track(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            call_log.append("generate")
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"fake-music-data")
-            return output_path
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            call_log.append("play")
-            proc = MagicMock()
-            proc.returncode = 0
-
-            async def _wait() -> int:
-                await asyncio.sleep(0.01)
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock()
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music.ElevenLabsMusicProvider"
-                    ".generate_track",
-                    fake_generate_track,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=Path("/tmp/vox-test-music"),
-                ),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-                await asyncio.sleep(0)
-
-                # Turn music on.
-                ctx.music_mode = "on"
-                ctx.music_owner = "test-session"
-                ctx.music_vibe = ("focused", "[calm]")
-                ctx.music_changed.set()
-                await asyncio.sleep(0.05)
-
-                # Turn music off.
-                ctx.music_mode = "off"
-                ctx.music_changed.set()
-                await asyncio.sleep(0.05)
-
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert "generate" in call_log
-        assert "play" in call_log
-
-    def test_crash_recovery_retries_with_backoff(self) -> None:
-        """Three failures in a row disable music mode."""
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "")
-        ctx.music_changed.set()
-
-        attempt_count = 0
-
-        async def failing_generate(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal attempt_count
-            attempt_count += 1
-            msg = f"generation failed (attempt {attempt_count})"
-            raise RuntimeError(msg)
-
-        async def _drive() -> None:
-            nonlocal attempt_count
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music.ElevenLabsMusicProvider"
-                    ".generate_track",
-                    failing_generate,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=Path("/tmp/vox-test-music"),
-                ),
-                patch("punt_vox.voxd._monolith._music_backoff_sleep", AsyncMock()),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-                # Yield control so the loop can run its 3 retries.
-                for _ in range(20):
-                    await asyncio.sleep(0)
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert attempt_count == 3
-        assert ctx.music_mode == "off"
-        assert ctx.music_state == "idle"
-
-    def test_vibe_change_during_generation_triggers_regeneration(self) -> None:
-        """Setting music_changed during generation causes a new track."""
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "")
-        ctx.music_changed.set()
-
-        generation_count = 0
-
-        async def counting_generate(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal generation_count
-            generation_count += 1
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"fake-music-data")
-
-            # On first generation, simulate a vibe change mid-flight.
-            if generation_count == 1:
-                ctx.music_vibe = ("happy", "[warm]")
-                ctx.music_changed.set()
-
-            return output_path
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            proc = MagicMock()
-            proc.returncode = 0
-
-            async def _wait() -> int:
-                await asyncio.sleep(0.01)
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock()
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music.ElevenLabsMusicProvider"
-                    ".generate_track",
-                    counting_generate,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=Path("/tmp/vox-test-music"),
-                ),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-                await asyncio.sleep(0.15)
-
-                ctx.music_mode = "off"
-                ctx.music_changed.set()
-                await asyncio.sleep(0.05)
-
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert generation_count >= 2
-
-
-class TestMusicLoopGaplessHandoff:
-    """Old track must keep looping while generation runs concurrently."""
-
-    def test_old_track_loops_during_generation(self) -> None:
-        """Playback subprocess stays alive the entire time generation runs.
-
-        Simulates a slow generation (~0.15s) and verifies the playback
-        subprocess is NOT killed until the new track is ready.
-        """
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "[calm]")
-        ctx.music_changed.set()
-
-        generation_count = 0
-        play_count = 0
-        # Track which procs were alive during generation.
-        procs_alive_during_gen: list[bool] = []
-
-        async def slow_generate(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal generation_count
-            generation_count += 1
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"fake-music-data")
-
-            if generation_count == 2:
-                # Second generation (triggered by vibe change). The old
-                # playback proc should still be alive during this window.
-                proc = ctx.music_proc
-                is_alive = proc is not None and proc.returncode is None
-                procs_alive_during_gen.append(is_alive)
-                # Simulate slow generation.
-                await asyncio.sleep(0.1)
-                # Check again after the sleep.
-                proc = ctx.music_proc
-                is_alive = proc is not None and proc.returncode is None
-                procs_alive_during_gen.append(is_alive)
-
-            return output_path
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            nonlocal play_count
-            play_count += 1
-            proc = MagicMock()
-            proc.returncode = None  # Still running.
-
-            async def _wait() -> int:
-                # Simulate a long track so it doesn't end naturally.
-                # Return immediately if the process was already killed,
-                # mirroring real OS behavior after SIGKILL.
-                if proc.returncode is not None:
-                    return int(proc.returncode)
-                await asyncio.sleep(5.0)
-                proc.returncode = 0
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock(side_effect=lambda: setattr(proc, "returncode", -9))
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music.ElevenLabsMusicProvider"
-                    ".generate_track",
-                    slow_generate,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=Path("/tmp/vox-test-handoff"),
-                ),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-                # Let initial generation + first playback start.
-                await asyncio.sleep(0.05)
-
-                # Trigger a vibe change while the first track is playing.
-                ctx.music_vibe = ("happy", "[warm]")
-                ctx.music_changed.set()
-
-                # Wait for second generation to complete + handoff.
-                await asyncio.sleep(0.3)
-
-                # Shut down.
-                ctx.music_mode = "off"
-                ctx.music_changed.set()
-                await asyncio.sleep(0.05)
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert generation_count >= 2, (
-            f"expected >=2 generations, got {generation_count}"
-        )
-        assert play_count >= 2, f"expected >=2 playback spawns, got {play_count}"
-        # The old playback proc was alive during the entire generation window.
-        assert all(procs_alive_during_gen), (
-            f"old track was killed during generation: {procs_alive_during_gen}"
-        )
-
-    def test_second_vibe_change_cancels_inflight_generation(self) -> None:
-        """A second vibe change during generation cancels the first and starts fresh."""
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "[calm]")
-        ctx.music_changed.set()
-
-        generation_vibes: list[str] = []
-        gen_event = asyncio.Event()
-
-        async def tracking_generate(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            vibe, _ = ctx.music_vibe
-            generation_vibes.append(vibe)
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"fake-music-data")
-
-            if len(generation_vibes) == 2:
-                # Signal that second generation started, then simulate slow work.
-                gen_event.set()
-                await asyncio.sleep(0.5)
-            elif len(generation_vibes) == 3:
-                # Third generation — the replacement after cancel.
-                pass
-
-            return output_path
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            proc = MagicMock()
-            proc.returncode = None
-
-            async def _wait() -> int:
-                if proc.returncode is not None:
-                    return int(proc.returncode)
-                await asyncio.sleep(5.0)
-                proc.returncode = 0
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock(side_effect=lambda: setattr(proc, "returncode", -9))
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music.ElevenLabsMusicProvider"
-                    ".generate_track",
-                    tracking_generate,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=Path("/tmp/vox-test-cancel"),
-                ),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-                await asyncio.sleep(0.05)
-
-                # First vibe change triggers generation #2.
-                ctx.music_vibe = ("happy", "[warm]")
-                ctx.music_changed.set()
-                # Wait for second generation to start.
-                await asyncio.wait_for(gen_event.wait(), timeout=1.0)
-
-                # Second vibe change while #2 is in-flight — should cancel it.
-                ctx.music_vibe = ("energetic", "[upbeat]")
-                ctx.music_changed.set()
-                await asyncio.sleep(0.3)
-
-                ctx.music_mode = "off"
-                ctx.music_changed.set()
-                await asyncio.sleep(0.05)
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert len(generation_vibes) >= 3, (
-            f"expected >=3 generation attempts, got "
-            f"{len(generation_vibes)}: {generation_vibes}"
-        )
-        # The third generation should have the latest vibe.
-        assert generation_vibes[-1] == "energetic"
-
-
-class TestKillMusicProc:
-    """_kill_music_proc safely terminates the music subprocess."""
-
-    def test_kills_running_proc(self) -> None:
-        ctx = _make_ctx()
-        proc = MagicMock()
-        proc.returncode = None
-        proc.kill = MagicMock()
-        proc.wait = AsyncMock(return_value=0)
-        ctx.music_proc = proc
-
-        asyncio.run(_kill_music_proc(ctx))
-
-        proc.kill.assert_called_once()
-        assert ctx.music_proc is None
-
-    def test_noop_when_no_proc(self) -> None:
-        ctx = _make_ctx()
-        ctx.music_proc = None
-
-        asyncio.run(_kill_music_proc(ctx))
-
-        assert ctx.music_proc is None
-
-    def test_noop_when_proc_already_exited(self) -> None:
-        ctx = _make_ctx()
-        proc = MagicMock()
-        proc.returncode = 0
-        proc.kill = MagicMock()
-        ctx.music_proc = proc
-
-        asyncio.run(_kill_music_proc(ctx))
-
-        proc.kill.assert_not_called()
-        assert ctx.music_proc is None
-
-
 class TestEmptyOwnerIdRejection:
     """Handlers must reject empty owner_id to prevent ownership spoofing."""
 
@@ -2588,64 +949,6 @@ class TestEmptyOwnerIdRejection:
             {"type": "error", "id": "empty-4", "message": "owner_id is required"}
         )
         assert ctx.music_vibe == ("", "")
-
-
-class TestMusicLoopLostWakeup:
-    """_music_loop must not block when music_mode is set before clear()."""
-
-    def test_mode_on_before_wait_skips_blocking(self) -> None:
-        """If music_mode becomes 'on' between clear() and wait(), proceed."""
-        ctx = _make_ctx()
-        # Pre-set music_mode to "on" so the re-check after clear() catches it.
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "[calm]")
-        # Do NOT set music_changed — the loop must detect mode via re-check.
-
-        generation_happened = False
-
-        async def fake_generate(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal generation_happened
-            generation_happened = True
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"fake")
-            # Turn off to let the loop exit cleanly.
-            ctx.music_mode = "off"
-            ctx.music_changed.set()
-            return output_path
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music.ElevenLabsMusicProvider"
-                    ".generate_track",
-                    fake_generate,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=Path("/tmp/vox-test-lost-wakeup"),
-                ),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-                # Give the loop enough time to either proceed or block.
-                await asyncio.sleep(0.1)
-                if not generation_happened:
-                    # Loop is stuck — cancel and fail.
-                    task.cancel()
-                    with contextlib.suppress(asyncio.CancelledError):
-                        await task
-                else:
-                    task.cancel()
-                    with contextlib.suppress(asyncio.CancelledError):
-                        await task
-
-        asyncio.run(_drive())
-
-        assert generation_happened, (
-            "music_loop blocked on wait despite music_mode=='on'"
-        )
 
 
 class TestAutoTrackName:
@@ -2968,347 +1271,3 @@ class TestHandlerRegistration:
 
         assert "music_list" in _HANDLERS
         assert _HANDLERS["music_list"] is _handle_music_list
-
-
-class TestGenFailureKeepsOldTrack:
-    """Generation failure must not kill the old track subprocess.
-
-    Covers the fix for issue vox-m2l: when the generation task fails
-    during the playback loop, the old track keeps looping during
-    retry/backoff. Only max-retries or a successful handoff kills it.
-    """
-
-    def test_failure_then_success_old_track_alive_throughout(
-        self, tmp_path: Path
-    ) -> None:
-        """First generation (vibe change) fails, retry succeeds.
-
-        The old playback subprocess must remain alive (returncode is
-        None) during the entire failure + backoff + retry window.
-        """
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "[calm]")
-        ctx.music_changed.set()
-
-        generation_count = 0
-        # Snapshots of old-proc liveness taken during the failing gen
-        # and during the retry gen.
-        old_proc_alive_snapshots: list[bool] = []
-
-        async def fail_then_succeed(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal generation_count
-            generation_count += 1
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-
-            if generation_count == 2:
-                # Second generation (triggered by vibe change): FAIL.
-                # Snapshot old proc liveness before raising.
-                proc = ctx.music_proc
-                old_proc_alive_snapshots.append(
-                    proc is not None and proc.returncode is None,
-                )
-                msg = "network error"
-                raise RuntimeError(msg)
-
-            if generation_count == 3:
-                # Third generation (retry after failure): succeed.
-                # The old proc should STILL be alive during retry.
-                proc = ctx.music_proc
-                old_proc_alive_snapshots.append(
-                    proc is not None and proc.returncode is None,
-                )
-
-            output_path.write_bytes(b"fake-music-data")
-            return output_path
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            proc = MagicMock()
-            proc.returncode = None
-
-            async def _wait() -> int:
-                if proc.returncode is not None:
-                    return int(proc.returncode)
-                await asyncio.sleep(5.0)
-                proc.returncode = 0
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock(
-                side_effect=lambda: setattr(proc, "returncode", -9),
-            )
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music."
-                    "ElevenLabsMusicProvider.generate_track",
-                    fail_then_succeed,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=tmp_path / "vox-test-gen-fail",
-                ),
-                patch("punt_vox.voxd._monolith._music_backoff_sleep", AsyncMock()),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-
-                # Poll until initial generation + first playback start.
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if ctx.music_proc is not None:
-                        break
-
-                # Trigger vibe change — second generation will fail.
-                ctx.music_vibe = ("happy", "[warm]")
-                ctx.music_changed.set()
-
-                # Poll until failure + backoff + retry + handoff complete.
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if generation_count >= 3:
-                        break
-
-                ctx.music_mode = "off"
-                ctx.music_changed.set()
-                for _ in range(50):
-                    await asyncio.sleep(0.01)
-                    if ctx.music_state == "idle":
-                        break
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        # Generation ran at least 3 times: initial, fail, retry.
-        assert generation_count >= 3, (
-            f"expected >=3 generations, got {generation_count}"
-        )
-        # Old track was alive during both the failing gen and the retry.
-        assert len(old_proc_alive_snapshots) >= 2, (
-            f"expected >=2 liveness snapshots, got {old_proc_alive_snapshots}"
-        )
-        assert all(old_proc_alive_snapshots), (
-            f"old track was killed during gen failure/retry: {old_proc_alive_snapshots}"
-        )
-
-    def test_max_retries_stops_music_mode(self, tmp_path: Path) -> None:
-        """After max retries during playback, music_mode becomes 'off'."""
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "[calm]")
-        ctx.music_changed.set()
-
-        generation_count = 0
-
-        async def always_fail_after_first(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal generation_count
-            generation_count += 1
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-
-            if generation_count == 1:
-                # Initial generation succeeds.
-                output_path.write_bytes(b"fake-music-data")
-                return output_path
-
-            # All subsequent generations fail.
-            msg = f"generation failed (attempt {generation_count})"
-            raise RuntimeError(msg)
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            proc = MagicMock()
-            proc.returncode = None
-
-            async def _wait() -> int:
-                if proc.returncode is not None:
-                    return int(proc.returncode)
-                await asyncio.sleep(5.0)
-                proc.returncode = 0
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock(
-                side_effect=lambda: setattr(proc, "returncode", -9),
-            )
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music."
-                    "ElevenLabsMusicProvider.generate_track",
-                    always_fail_after_first,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=tmp_path / "vox-test-gen-maxretry",
-                ),
-                patch("punt_vox.voxd._monolith._music_backoff_sleep", AsyncMock()),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-
-                # Poll until initial generation + first playback start.
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if ctx.music_proc is not None:
-                        break
-
-                # Trigger vibe change — all subsequent gens will fail.
-                ctx.music_vibe = ("happy", "[warm]")
-                ctx.music_changed.set()
-
-                # Poll for 3 failures + final shutdown.
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if ctx.music_mode == "off":
-                        break
-
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert ctx.music_mode == "off"
-        assert ctx.music_state == "idle"
-        # 1 initial success + 3 failures = 4 total.
-        assert generation_count == 4, (
-            f"expected 4 generations (1 ok + 3 fail), got {generation_count}"
-        )
-
-    def test_successful_handoff_after_retry_resets_counter(
-        self, tmp_path: Path
-    ) -> None:
-        """A successful handoff after one failure resets the retry counter."""
-        ctx = _make_ctx()
-        ctx.music_mode = "on"
-        ctx.music_owner = "test-session"
-        ctx.music_vibe = ("focused", "[calm]")
-        ctx.music_changed.set()
-
-        generation_count = 0
-
-        async def fail_once_per_cycle(
-            self: object, prompt: str, duration_ms: int, output_path: Path
-        ) -> Path:
-            nonlocal generation_count
-            generation_count += 1
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-
-            # Fail once per vibe-change cycle: gen #2 and gen #4.
-            if generation_count in (2, 4):
-                msg = "transient error"
-                raise RuntimeError(msg)
-
-            output_path.write_bytes(b"fake-music-data")
-            return output_path
-
-        async def fake_subprocess_exec(*args: object, **kwargs: object) -> MagicMock:
-            proc = MagicMock()
-            proc.returncode = None
-
-            async def _wait() -> int:
-                if proc.returncode is not None:
-                    return int(proc.returncode)
-                await asyncio.sleep(5.0)
-                proc.returncode = 0
-                return 0
-
-            proc.wait = _wait
-            proc.kill = MagicMock(
-                side_effect=lambda: setattr(proc, "returncode", -9),
-            )
-            return proc
-
-        async def _drive() -> None:
-            with (
-                patch(
-                    "punt_vox.providers.elevenlabs_music."
-                    "ElevenLabsMusicProvider.generate_track",
-                    fail_once_per_cycle,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith.asyncio.create_subprocess_exec",
-                    fake_subprocess_exec,
-                ),
-                patch(
-                    "punt_vox.voxd._monolith._music_output_dir",
-                    return_value=tmp_path / "vox-test-gen-reset",
-                ),
-                patch("punt_vox.voxd._monolith._music_backoff_sleep", AsyncMock()),
-            ):
-                task = asyncio.create_task(_music_loop(ctx))
-
-                # Poll until initial generation + first playback start.
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if ctx.music_proc is not None:
-                        break
-
-                # Trigger vibe change — gen #2 fails, #3 succeeds.
-                ctx.music_vibe = ("happy", "[warm]")
-                ctx.music_changed.set()
-
-                # Poll until retry succeeds (generation_count >= 3).
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if generation_count >= 3:
-                        break
-
-                # Music should still be on — the retry succeeded.
-                assert ctx.music_mode == "on"
-
-                # Now trigger a SECOND vibe change to prove the retry
-                # counter was actually reset. gen #4 will fail, #5
-                # will succeed. If the counter had accumulated from
-                # the first failure cycle, this second failure would
-                # push past max retries and turn music off.
-                ctx.music_vibe = ("energetic", "[bold]")
-                ctx.music_changed.set()
-
-                for _ in range(100):
-                    await asyncio.sleep(0.01)
-                    if generation_count >= 5:
-                        break
-
-                # Music is STILL on — the counter was reset by the
-                # first successful handoff and the second cycle's
-                # single failure did not exceed max retries.
-                assert ctx.music_mode == "on", (
-                    "retry counter was not reset: second failure cycle "
-                    "pushed past max retries"
-                )
-
-                ctx.music_mode = "off"
-                ctx.music_changed.set()
-                for _ in range(50):
-                    await asyncio.sleep(0.01)
-                    if ctx.music_state == "idle":
-                        break
-                task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await task
-
-        asyncio.run(_drive())
-
-        assert generation_count >= 5, (
-            f"expected >=5 generations (2 cycles of fail+succeed), "
-            f"got {generation_count}"
-        )
-        # Music stayed on through both failure+retry cycles.
